+++ src/utils/exceptionThrowers.ts
@@ -17,6 +17,20 @@ import ChatSDKExceptionDetails from "../core/ChatSDKExceptionDetails";
 import ScenarioMarker from "../telemetry/ScenarioMarker";
 import TelemetryEvent from "../telemetry/TelemetryEvent";
 
+class CustomChatSDKError {
+    public message: string;
+    public httpResponseStatusCode: number | undefined;
+
+    constructor(message: string, httpResponseStatusCode?: number) {
+        this.message = message;
+        this.httpResponseStatusCode = httpResponseStatusCode;
+    }
+
+    toString(): string {
+        return this.message;
+    }
+}
+
 export const throwChatSDKError = (chatSDKError: ChatSDKErrors, e: unknown, scenarioMarker: ScenarioMarker, telemetryEvent: TelemetryEvent, telemetryData: {[key: string]: string} = {}, message = ""): void => {
     const exceptionDetails: ChatSDKExceptionDetails = {
         response: chatSDKError
@@ -36,7 +50,11 @@ export const throwChatSDKError = (chatSDKError: ChatSDKErrors, e: unknown, scena
         console.error(message);
     }
 
-    throw new Error(exceptionDetails.response);
+    throw new CustomChatSDKError(
+        exceptionDetails.response,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        ((e as any)?.isAxiosError && (e as any)?.response?.status) ? (e as any)?.response?.status : undefined
+    );
 }
 
 export const throwScriptLoadFailure = (e: unknown, scenarioMarker: ScenarioMarker, telemetryEvent: TelemetryEvent): void => {